+++ include/OBDHandler.hpp
@@ -29,7 +29,7 @@ private:
     void updateAll();
     void printAll() const;
     void addNewInfo(uint16_t pid, uint8_t length, std::string description, double current,
-                    double min, double max, double increment, uint64_t pace,
+                    double min, double max, double increment, uint32_t pace,
                     std::function<int32_t(const double &)> customGetter);
 };
 #endif // OBD_HANDLER_HPP
