+++ src/OBDHandler.cpp
@@ -84,7 +84,7 @@ const OBDInfo *OBDHandler::getByIndex(uint8_t index) const
 }
 
 void OBDHandler::addNewInfo(uint16_t pid, uint8_t length, std::string description, double current,
-                            double min, double max, double increment, uint64_t pace,
+                            double min, double max, double increment, uint32_t pace,
                             std::function<int32_t(const double &)> customGetter)
 {
     m_infos.push_back(OBDInfo(pid, length, description, current, min, max, increment, pace, customGetter));
